fix: add cluster
