+++ crud/app.js
@@ -1,0 +1,3 @@
+import createCRUD from './index.js'
+import dotenv from 'dotenv'
+createCRUD(dotenv.config().parsed.PORT)
