+++ crud/db.js
@@ -1,0 +1,15 @@
+import { v4 as uuidv4 } from 'uuid'; 
+export let users = [
+  {
+    id: uuidv4(),
+    username: 'misha',
+    age: 25,
+    hobbies : []
+  },
+  {
+    id: uuidv4(),
+    username: 'oleg',
+    age: 18,
+    hobbies : []
+  }
+]
